error to wasm buffer.
